Realized that video publishing was already implemented in old commits of video stream, publishing locally now.
