+++ src/radio/radio/receiver_udp.py
@@ -33,8 +33,6 @@ class UDPReceiver(Node):
         
         if data_type == "KEY":
             topic = "thruster_control"
-        elif data_type == "VIDEO":
-            topic = "video_stream"
         elif data_type == "RADAR":
             topic = "radar_data"
         else: #add other keyboard or controller commands
